+++ shared/src/simulation/activities/reassign-treatments.ts
@@ -16,7 +16,6 @@ import {
 import type { Immutable, Mutable } from '../../utils/immutability';
 import { UUID, uuidValidationOptions } from '../../utils/uuid';
 import { IsLiteralUnion, IsValue } from '../../utils/validators';
-// Do not import from "../utils" since it would cause circular dependencies
 import {
     TreatmentProgress,
     treatmentProgressAllowedValues,
@@ -131,7 +130,9 @@ export const reassignTreatmentsActivity: SimulationActivity<ReassignTreatmentsAc
             }
 
             let allowTerminate = true;
-            let missingPersonnel: PersonnelResource | null = null;
+            let missingPersonnel:
+                | ResourceDescription<PersonnelType>
+                | undefined;
 
             const progress = activityState.treatmentProgress;
             switch (progress) {
@@ -201,16 +202,21 @@ export const reassignTreatmentsActivity: SimulationActivity<ReassignTreatmentsAc
                 }
             }
 
-            if (missingPersonnel && !isEmptyResource(missingPersonnel)) {
-                sendSimulationEvent(
-                    simulatedRegion,
-                    ResourceRequiredEvent.create(
-                        nextUUID(draftState),
-                        simulatedRegion.id,
-                        missingPersonnel,
-                        'reassignTreatmentsActivity'
-                    )
+            if (missingPersonnel !== undefined) {
+                const missingResource = PersonnelResource.create(
+                    ceilResourceDescription(missingPersonnel)
                 );
+                if (!isEmptyResource(missingResource)) {
+                    sendSimulationEvent(
+                        simulatedRegion,
+                        ResourceRequiredEvent.create(
+                            nextUUID(draftState),
+                            simulatedRegion.id,
+                            missingResource,
+                            'reassignTreatmentsActivity'
+                        )
+                    );
+                }
             }
 
             if (allowTerminate) {
@@ -246,6 +252,11 @@ type PersonnelToPatientCategoryDict = {
         };
     };
 };
+
+interface PersonnelSubstitution {
+    from: ResourceDescription<PersonnelType>;
+    to: Mutable<PersonnelType>;
+}
 
 function createCateringMaterials(
     materials: Mutable<Material>[]
@@ -382,11 +393,13 @@ function triage(
     patients: Mutable<Patient>[],
     personnel: Mutable<Personnel>[],
     materials: Mutable<Material>[]
-): [boolean, PersonnelResource] {
+): [boolean, ResourceDescription<PersonnelType>] {
     const patientsToTreat: Mutable<Patient>[] = [];
     const cateringPersonnel = createCateringPersonnel(personnel).sort(
         (a, b) => a.priority - b.priority
     );
+    const emptyPersonnelCount = PersonnelResource.create().personnelCounts;
+    const triagePersonnelSubstitution: PersonnelSubstitution[] = [];
 
     patients.forEach((patient) => {
         if (Patient.pretriageStatusIsLocked(patient)) {
@@ -408,7 +421,15 @@ function triage(
 
         if (triagePersonnelIndex !== -1) {
             // Personnel that is used for triage shall not be used for treatments
-            cateringPersonnel.splice(triagePersonnelIndex, 1);
+            const triagePersonnel = cateringPersonnel.splice(
+                triagePersonnelIndex,
+                1
+            );
+            // But we can get it back
+            triagePersonnelSubstitution.push({
+                to: triagePersonnel[0]!.personnel.personnelType,
+                from: emptyPersonnelCount,
+            });
         }
     });
 
@@ -416,10 +437,23 @@ function triage(
         patients.length - patientsToTreat.length,
         personnel
     );
+    let [, treatmentMissingPersonnel] = assignTreatments(
+        draftState,
+        patientsToTreat,
+        cateringPersonnel,
+        materials
+    );
 
-    assignTreatments(draftState, patientsToTreat, cateringPersonnel, materials);
+    treatmentMissingPersonnel = applySubstitutions(
+        treatmentMissingPersonnel,
+        triagePersonnelSubstitution
+    );
+
     const triageDone = patientsToTreat.length === patients.length;
-    return [triageDone, missingPersonnel];
+    return [
+        triageDone,
+        addResourceDescription(missingPersonnel, treatmentMissingPersonnel),
+    ];
 }
 
 /**
@@ -436,7 +470,7 @@ function treat(
     patients: Mutable<Patient>[],
     personnel: Mutable<Personnel>[],
     materials: Mutable<Material>[]
-): [boolean, PersonnelResource] {
+): [boolean, ResourceDescription<PersonnelType>] {
     return assignTreatments(
         draftState,
         patients,
@@ -598,7 +632,7 @@ function assignTreatments(
     patients: Mutable<Patient>[],
     cateringPersonnel: CateringPersonnel[],
     materials: Mutable<Material>[]
-): [boolean, PersonnelResource] {
+): [boolean, ResourceDescription<PersonnelType>] {
     const groupedPatients = groupBy(patients, (patient) =>
         Patient.getVisibleStatus(
             patient,
@@ -700,7 +734,7 @@ function assignTreatments(
         personnelTreatments,
         patientTreatments
     );
-    return [secured, PersonnelResource.create(missingPersonnel)];
+    return [secured, missingPersonnel];
 }
 
 function tryAssignPersonnel(
@@ -777,7 +811,7 @@ function recordPatientTreatedByPersonnel(
 function estimateRequiredPersonnel(
     patientCount: number,
     personnel: Personnel[]
-): PersonnelResource {
+): ResourceDescription<PersonnelType> {
     const groupedPersonnel = groupBy(personnel, (p) => p.personnelType);
     const havePersonnel = Object.fromEntries(
         StrictObject.keys(personnelPriorities).map((pt) => [
@@ -789,12 +823,10 @@ function estimateRequiredPersonnel(
         scaleResourceDescription(estimatedSkDistribution, patientCount)
     );
 
-    // We want the missing personnel exactly, no substitution of personnel tiers takes place.
-    return PersonnelResource.create(
-        maxResourceDescription(
-            subtractResourceDescription(wantPersonnel, havePersonnel),
-            0
-        )
+    // We want the missing personnel exactly, no substitution of personnel qualification takes place.
+    return maxResourceDescription(
+        subtractResourceDescription(wantPersonnel, havePersonnel),
+        0
     );
 }
 
@@ -813,20 +845,16 @@ function requiredPersonnelForPatients(
         )
     );
     const result = addResourceDescription(
-        ceilResourceDescription(
-            addResourceDescription(
-                requiredByType['red']!,
-                requiredByType['yellow']!
-            )
+        addResourceDescription(
+            requiredByType['red']!,
+            requiredByType['yellow']!
         ),
         // Green should not be mixed. Therefore we ceil them before merging.
         ceilResourceDescription(requiredByType['green']!)
     );
-    result.notarzt = Math.ceil(
-        Math.max(
-            requiredByType['red']!.notarzt,
-            requiredByType['yellow']!.notarzt
-        )
+    result.notarzt = Math.max(
+        requiredByType['red']!.notarzt,
+        requiredByType['yellow']!.notarzt
     );
     return result;
 }
@@ -882,11 +910,9 @@ function calculateMissingPersonnel(
     ) as { [K in TreatablePatientStatus]: ResourceDescription<PersonnelType> };
 
     const totalMissingPersonnel = addResourceDescription(
-        ceilResourceDescription(
-            addResourceDescription(
-                patientStatusMissingPersonnel.red,
-                patientStatusMissingPersonnel.yellow
-            )
+        addResourceDescription(
+            patientStatusMissingPersonnel.red,
+            patientStatusMissingPersonnel.yellow
         ),
         ceilResourceDescription(patientStatusMissingPersonnel.green)
     );
@@ -919,7 +945,6 @@ function calculateSubstitutedMissingPersonnel(
     cateringDict: { [k: string]: CateringPersonnel },
     personnelTreatments: PersonnelToPatientCategoryDict
 ): ResourceDescription<PersonnelType> {
-    let missing = { ...missingPersonnel };
     const substitutions = StrictObject.entries(personnelTreatments)
         .map(
             ([persId, roles]) =>
@@ -950,17 +975,28 @@ function calculateSubstitutedMissingPersonnel(
             return { from, to: realPersonnelType };
         });
 
+    return applySubstitutions(missingPersonnel, substitutions);
+}
+
+function applySubstitutions(
+    missing: ResourceDescription<PersonnelType>,
+    substitutions: PersonnelSubstitution[]
+) {
+    let stillMissing = { ...missing };
     reversedPersonnelTypePriorityList.forEach((personnelType) => {
-        while (missing[personnelType] > 0) {
+        while (stillMissing[personnelType] > 0) {
             const substitutionIndex = substitutions.findIndex(
                 (s) => s.to === personnelType
             );
             if (substitutionIndex === -1) break;
             const substitution = substitutions.splice(substitutionIndex, 1)[0]!;
-            missing[substitution!.to] -= 1;
+            stillMissing[substitution!.to] -= 1;
 
-            missing = addResourceDescription(missing, substitution.from);
+            stillMissing = addResourceDescription(
+                stillMissing,
+                substitution.from
+            );
         }
     });
-    return maxResourceDescription(ceilResourceDescription(missing), 0);
+    return maxResourceDescription(stillMissing, 0);
 }