+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
 -   When personnel is missing during patient treatment in a simulated region, the reassign treatment activity now sends an event to notify the region about the shortage.
 -   A new behavior has been added to respond to personnel shortages by instructing the region to request new vehicles.
     -   The priorities of vehicles to request can be configured in a new behavior tab.
+-   New patients added to simulated regions during treatment are triaged and treated and the phase can change back if more personnel is required.
 
 ### Changed